pyproject.toml
[project]
name = "bi-agent"
version = "2.0.0"
description = "Enterprise BI Agent — voice + text to SQL, chart, insight, and explanation"
requires-python = ">=3.11"

dependencies = [
    # ── Web framework ─────────────────────────────────────────
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",

    # ── Database ──────────────────────────────────────────────
    "pyodbc>=5.1.0",
    "sqlalchemy>=2.0.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",

    # ── Google AI (ADK + Gemini) ──────────────────────────────
    "google-adk>=0.3.0",
    "google-genai>=0.8.0",

    # ── Charting ──────────────────────────────────────────────
    "altair>=5.3.0",

    # ── Utilities ─────────────────────────────────────────────
    "pydantic>=2.7.0",
]

[project.scripts]
bi-agent = "main:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main", "pipeline", "tools", "db", "chart_builder"]

[tool.setuptools.packages.find]
include = ["agents*"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]   # pycodestyle errors, pyflakes, isort
ignore = ["E501"]           # line length handled by formatter

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]